requirements/base.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.7
django-cors-headers==3.13.0
django-environ==0.9.0
django-phonenumber-field==7.0.2
django-rest-framework==0.1.0
django-twilio==0.14.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
install==1.3.5
phonenumbers==8.13.6
Pillow==9.4.0
PyJWT==2.6.0
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
twilio==7.16.3
urllib3==1.26.14
whitenoise==6.2.0  # https://github.com/evansd/whitenoise
django-filter==22.1
drf-spectacular==0.23.1  # https://github.com/tfranzel/drf-spectacular
